requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
djangorestframework==3.14.0
dnspython==2.3.0
idna==3.4
mongomock==4.1.2
packaging==23.0
pymongo==4.3.3
pytz==2022.7.1
requests==2.28.2
sentinels==1.0.0
sqlparse==0.4.3
urllib3==1.26.14
